load vesting data from file in VestingManager and move test [skip ci]
