+++ massa-execution-worker/src/tests/mod.rs
@@ -22,3 +22,6 @@ mod interface;
     feature = "testing"
 ))]
 pub use mock::get_sample_state;
+
+#[cfg(any(test, feature = "testing"))]
+mod tests_vesting_manager;
